refactor: split AI tests into separate files as requested
- Created `MinimaxTests.cs`
- Created `AlphaBetaTests.cs`
- Kept `EvaluationTests.cs`
- Removed `AISearchTests.cs`
- Updated documentation links if necessary
